Separate custom widgets into a different file
