+++ lib/pages/authentication_page/authentication_widgets/profile_setup_pages.dart
@@ -1,5 +1,5 @@
+import 'package:flilipino_food_app/pages/authentication_page/authentication_widgets/colored_inputs.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 
 class UserDetails extends StatelessWidget {
   const UserDetails({super.key});
@@ -36,62 +36,6 @@ class UserDetails extends StatelessWidget {
           hintText: "dd / mm / yyyy",
         ),
       ],
-    );
-  }
-}
-
-class ColoredInputNumber extends StatelessWidget {
-  const ColoredInputNumber({
-    super.key,
-    this.controller,
-    this.suffixText,
-    required this.hintText,
-  });
-
-  final String hintText;
-  final String? suffixText;
-  final TextEditingController? controller;
-
-  @override
-  Widget build(BuildContext context) {
-    return TextFormField(
-      controller: controller,
-      decoration: InputDecoration(
-        border: OutlineInputBorder(
-          borderSide: BorderSide.none,
-          borderRadius: BorderRadius.circular(16),
-        ),
-        filled: true,
-        hintText: hintText,
-        suffixText: suffixText,
-        fillColor: Theme.of(context).colorScheme.secondaryContainer,
-      ),
-      keyboardType: TextInputType.number,
-      inputFormatters: [LengthLimitingTextInputFormatter(4)],
-    );
-  }
-}
-
-// Something to place, does not work yet
-class ColoredPlaceholder extends StatelessWidget {
-  const ColoredPlaceholder({
-    super.key,
-    required this.hintText,
-  });
-  final String hintText;
-
-  @override
-  Widget build(BuildContext context) {
-    return TextFormField(
-      decoration: InputDecoration(
-        border: OutlineInputBorder(
-          borderSide: BorderSide.none,
-          borderRadius: BorderRadius.circular(16),
-        ),
-        filled: true,
-        hintText: hintText,
-        fillColor: Theme.of(context).colorScheme.secondaryContainer,
-      ),
     );
   }
 }
